ZCS-5491:browseBy="objects" returns no hits
As a result of zimlets being disabled, BrowseRequest with
browseBy="objects" will always return an empty list.  This is because
indexing information is only added if any object handlers find matches
in the text and object handlers come from zimlets, so if there are no
zimlets, there are no object handlers.