ZCS-7951: Fix for tgz Export/Import test cases
This is originating because of the following commit 5406034
in the zm-mailbox repo for the work that has been done in ZCS-7700 Override zimbra_ssl and
zimbra_admin upstreams. So, on removing the above commit for ZCS-7700 the test was passing.
In the newer changes the binary data is not getting saved to the file. In the test cases on
changing the format from tgz to tar and all the tests of Export/Import seem to be passing.
After the change of the regular and admin SOAP requests to pass-through, the new endpoint
tgz format doesn't seem to be working but on changing it to tar it is working.

So, on seeing further into the RestServletTest class file noticed that in processResponseHeaders
method only we are processing the headers if it has Transfer-Encoding or it's Content-Type is text
or application/zip but not for the application/x-compressed-tar. This was the reason on changing the
compression scheme from Gzip to BZip2 Compression scheme in the TgzFormatter class the header was
having Transfer-Encoding as chunked in the halft of the test cases which were passing and rest were
failing. So, to fix that also processing the header's Content-Type as application/x-compressed-tar.