Feature/myzimbra x (#100)
* Provision made to ignore tests

* Build.xml changed to ignore staf tests

* Added global.properties.in file

* Replaced values with env variables

* Changed path for test case lists

* ZCS:7179 Adding variable for retry

* ZCS:7179 Provided mechanism to do polling of indexed data for Search operations in SOAP Harness

* Removed duplicate entries from file as well as added new entries to file which will skip from next soap run

* ZCS-7488: Adding automation test case for zimbraSAN

* ZCS-7488: fix pr feedback

* Avoiding hardcoding of path for sanity test folder

* Avoiding hardcoding of path for sanity test folder

* ZCS-7609: Soap automation analysis for Passed on Rerun/New Failure tests (#80)

* ZCS-7609: update fix for few test cases

* ZCS-7609: updating fixes so far

* ZCS-7609: updating fixes so far

* ZCS-7609: updating fix for 1 soap test case

* ZCS-7609: updating soap harness config with mailbox hostname value and related test cases

* ZCS-7609: fix for zimbra Mail Host

* ZCS-7788: Create an LDAP attribute to store account Lockout reason (#84)

* ZCS-7788: Create an LDAP attribute to store account Lockout reason

* ClientInfoRequest api not returning response in JSON format

PR comments

Removed comments

* ZCS-7951: Fix for tgz Export/Import test cases

This is originating because of the following commit 5406034
in the zm-mailbox repo for the work that has been done in ZCS-7700 Override zimbra_ssl and
zimbra_admin upstreams. So, on removing the above commit for ZCS-7700 the test was passing.
In the newer changes the binary data is not getting saved to the file. In the test cases on
changing the format from tgz to tar and all the tests of Export/Import seem to be passing.
After the change of the regular and admin SOAP requests to pass-through, the new endpoint
tgz format doesn't seem to be working but on changing it to tar it is working.

So, on seeing further into the RestServletTest class file noticed that in processResponseHeaders
method only we are processing the headers if it has Transfer-Encoding or it's Content-Type is text
or application/zip but not for the application/x-compressed-tar. This was the reason on changing the
compression scheme from Gzip to BZip2 Compression scheme in the TgzFormatter class the header was
having Transfer-Encoding as chunked in the halft of the test cases which were passing and rest were
failing. So, to fix that also processing the header's Content-Type as application/x-compressed-tar.

* Framework change for handling ews delays

* Build.xml changes

* Pushing properties changes

* Changing testRoot path

* Review comments addressed

* PR comments

* Added zimbraX specific skip tests as well as properties file and added polling interval for Get related responses

* Done changes suggested by Swapnil in Previous PR

* Updated build.xml with uncommented if condition which were not needed for zimbrax

* Updated conf/ListOfTestCases.txt as per develop branch

Co-authored-by: SYNBUF\swapnil.pingle <swapnil.pingle@synacor.com>
Co-authored-by: Ajinkya Nahar <ajinkya.nahar@synacor.com>
Co-authored-by: log2akshat <akshat-pg8@iiitmk.ac.in>
Co-authored-by: girish.bhamare <girish.bhamare@syancor.com>